+++ tests/test_balinese.py
@@ -40,6 +40,7 @@ CONVERSION_CASES = {
     NumeralSystem.BALINESE: BALINESE_DIGITS,
     NumeralSystem.NEW_TAI_LUE: "᧐᧑᧒᧓᧔᧕᧖᧗᧘᧙",
     NumeralSystem.SAURASHTRA: "꣐꣑꣒꣓꣔꣕꣖꣗꣘꣙",
+    NumeralSystem.JAVANESE: "꧐꧑꧒꧓꧔꧕꧖꧗꧘꧙",
 }